add settings.bundle and remove settings view
