+++ Apollon/Custom/Extensions/DateExtension.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+extension Date {
+    func ISO8601FormatWithFractionalSeconds() -> String {
+        let formatter = DateFormatter()
+        formatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSSZ"
+        return formatter.string(from: self)
+    }
+}
