+++ Apollon/Models/ApollonDiagram.swift
@@ -18,11 +18,3 @@ import ApollonShared
         self.model = model ?? ""
     }
 }
-
-extension Date {
-    func ISO8601FormatWithFractionalSeconds() -> String {
-        let formatter = DateFormatter()
-        formatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSSZ"
-        return formatter.string(from: self)
-    }
-}
